Now features output of JSON file at private static String OUTPUT_PATHNAME in FormatOutput.java. If the file doesn't exist at the location, it makes a file in that directory with that name. Otherwise, it updates the existing file with a new JSONArray of vehicles keyed by timestamp. Currently pretty raw data for each vehicle (basically vehicleId, routeId, tripId, latitude, longitude, and bearing) for each vehicle. So the next step will be creating more usable data, such as distance to next stop, the nextStopName, previousStopName, etc.
Efficiency of the update will have to be considered. At the moment the
file gets read in line by line by a buffered reader.

Lastly, I included the Apache 2.0 license text in every file as is
recommended when using the file.

Meanwhile the test.json file can be loaded into a dict object in